USE="opensync" with plugins for both 0.2x and 0.3x series.
Package-Manager: portage-2.2.0_alpha28/cvs/Linux x86_64

+++ app-pda/synce-sync-engine/synce-sync-engine-0.15.1-r1.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2011 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/app-pda/synce-sync-engine/synce-sync-engine-0.15.1-r1.ebuild,v 1.3 2011/02/25 22:13:34 ssuominen Exp $
+# $Header: /var/cvsroot/gentoo-x86/app-pda/synce-sync-engine/synce-sync-engine-0.15.1-r1.ebuild,v 1.4 2011/03/28 01:47:32 ssuominen Exp $
 
 EAPI=3
 
@@ -8,7 +8,7 @@ PYTHON_DEPEND="2:2.6"
 SUPPORT_PYTHON_ABIS="1"
 RESTRICT_PYTHON_ABIS="3.*"
 
-inherit distutils
+inherit distutils multilib
 
 DESCRIPTION="A synchronization engine for SynCE"
 HOMEPAGE="http://www.synce.org/"
@@ -17,7 +17,7 @@ SRC_URI="mirror://sourceforge/synce/${P}.tar.gz"
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="~amd64 ~x86"
-IUSE=""
+IUSE="opensync"
 
 RDEPEND="dev-libs/librapi2[python]
 	dev-libs/librra[python]
@@ -25,7 +25,8 @@ RDEPEND="dev-libs/librapi2[python]
 	dev-libs/libxml2[python]
 	dev-libs/libxslt[python]
 	dev-python/dbus-python
-	dev-python/pygobject"
+	dev-python/pygobject
+	opensync? ( >=app-pda/libopensync-0.22 )"
 DEPEND="${RDEPEND}"
 
 PYTHON_MODNAME="SyncEngine"
@@ -45,18 +46,22 @@ src_install() {
 
 	distutils_src_install
 
-	# opensync plug-in begin
+	### opensync plug-in BEGIN
 	find "${D}" -type d -name plugins -exec rm -rf {} +
 
-	docinto synce-opensync-plugin
-	dodoc plugins/synce-opensync-plugin-* || die
-	# end
+	if use opensync; then
+		insinto /usr/$(get_libdir)/opensync/python-plugin
+		local plug=plugins/synce-opensync-plugin-
+
+		if has_version ">=app-pda/libopensync-0.30"; then
+			newins ${plug}3x.py synce-plugin.py || die
+		else
+			newins ${plug}2x.py synce-plugin.py || die
+		fi
+
+		dodoc ${plug}3x.README || die
+	fi
+	### opensync plug-in END
 
 	rm -rf "${D}"/usr/foobar
-}
-
-pkg_postinst() {
-	distutils_pkg_postinst
-
-	einfo "See /usr/share/doc/${PF}/synce-opensync-plugin for opensync plug-in."
 }
